fixing json
